+++ packages/image_picker/image_picker_ios/example/ios/RunnerTests/MetaDataUtilTests.swift
@@ -8,6 +8,10 @@ import UIKit
 
 @testable import image_picker_ios
 
+#if canImport(image_picker_ios_objc)
+  @testable import image_picker_ios_objc
+#endif
+
 @Suite
 struct MetaDataUtilTests {
   @Test func getImageMIMETypeFromImageData() {