+++ packages/image_picker/image_picker_ios/ios/image_picker_ios/Package.swift
@@ -12,20 +12,26 @@ let package = Package(
     .iOS("13.0")
   ],
   products: [
-    .library(name: "image-picker-ios", targets: ["image_picker_ios"])
+    .library(name: "image-picker-ios", targets: ["image_picker_ios", "image_picker_ios_objc"])
   ],
   dependencies: [],
   targets: [
     .target(
       name: "image_picker_ios",
-      dependencies: [],
-      exclude: ["include/image_picker_ios-umbrella.h", "include/ImagePickerPlugin.modulemap"],
+      dependencies: [
+        "image_picker_ios_objc"
+      ],
       resources: [
         .process("Resources")
-      ],
+      ]
+    ),
+    .target(
+      name: "image_picker_ios_objc",
+      dependencies: [],
+      exclude: ["include/image_picker_ios-umbrella.h", "include/ImagePickerPlugin.modulemap"],
       cSettings: [
         .headerSearchPath("include/image_picker_ios")
       ]
-    )
+    ),
   ]
 )